serial: 8250_dw: Add skip_clk_set_rate quirk
On the StarFive JH7100 SoC the UART core clocks can't be set to exactly
16 * 115200Hz and many other common bitrates. Trying this will only
result in a higher input clock, but low enough that the UART's internal
divisor can't come close enough to the baud rate target. So rather than
try to set the input clock it's better to rely solely on the UART's
internal divisor.

Signed-off-by: Emil Renner Berthing <kernel@esmil.dk>