arch, mm: make kernel_page_present() always available
For architectures that enable ARCH_HAS_SET_MEMORY having the ability to
verify that a page is mapped in the kernel direct map can be useful
regardless of hibernation.

Add RISC-V implementation of kernel_page_present(), update its forward
declarations and stubs to be a part of set_memory API and remove ugly
ifdefery in inlcude/linux/mm.h around current declarations of
kernel_page_present().

The patch is modify to support RISCV only. The patch is origin from:
commit 32a0de886eb3 ("arch, mm: make kernel_page_present() always
available")

Signed-off-by: Sia Jee Heng <jeeheng.sia@starfivetech.com>