riscv: dts: Set specific cpu frequency when suspending
As JH7110 saves & restores all clock registers in hibernation.
The cpu voltage may not correct after restoring cpu freqency
from hibernation resume process.

In case it's suspended in 1.5GHz with 1.04v,
and it schedules to 750Mhz with 0.8v before resuming,
then the cpu clock will be restored to 1.5GHz, but the cpu
voltage still remains in 0.8v, then the system crashes.

So we set the cpu frequency to 750MHz, and it will keep
the cpu voltage matched when doing hibernation resuming.

Signed-off-by: Mason Huo <mason.huo@starfivetech.com>